+++ lib/aixm/component/runway.rb
@@ -277,7 +277,7 @@ module AIXM
         # @return [AIXM::A] magnetic orientation (magnetic bearing) in degrees
         def magnetic_orientation
           if geographic_orientation && runway.airport.declination
-            geographic_orientation + runway.airport.declination
+            geographic_orientation - runway.airport.declination
           end
         end